Move containers to appropriate folder
