+++ src/components/Header/Header.scss
@@ -4,7 +4,6 @@
     display: flex;
     justify-content: space-between;
     background-color: #e6f5f7;
-    // border-bottom: 4.5px solid #4e4e4e;
     padding: 5px;
 }
 
@@ -35,14 +34,4 @@
     margin-bottom: -18px;
     color: tomato;
     font-size: 30px;
-}
-
-.subtitle {
-    font-size: 20px;
-    font-style: italic;
-    margin-top: 55px;
-    margin-right: 10px;
-    display: relative;
-    right: 0;
-    color: #4e4e4e;
 }
