+++ src/components/ParentSection/ParentSection.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import PaletteSection from '../../components/PaletteSection/PaletteSection';
-import ProjectSection from '../../components/ProjectSection/ProjectSection';
+import ProjectSection from '../../containers/ProjectSection/ProjectSection';
 
 export const ParentSection = () => {
     return (
