+++ src/containers/PaletteControls/PaletteControls.scss
