+++ src/containers/PaletteControls/PaletteControls.test.js
